requirements.txt
streamlit==1.46.0
pandas==2.3.0
gspread==6.0.2
oauth2client==4.1.3
google-api-python-client==2.128.0
Pillow==11.2.1
openpyxl==3.1.2